Add files via upload
To log in to your account use the login.php code. To reset your password use the reset.php page to register a new account use the register.php page. The index.php page is where you will be directed to if you log in also you can log out of your account from this page. The verifydata.php contains the main PHP code to make all the functionalities work.